Ignores plugin install errors
Ignoring errors on plugin install since (for ansible < 2.0) we do not
know in advance if the plugin is already installed.
So, at install time, taks will fal if the plugin is already present.
Sure, we could pre-check and act accordingly, but since plugin is a
list, this would require hellish hacks.